requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
PyPDF2
anthropic
python-dotenv
html2text